[IMP] l10n_de_reports: taxes
We want to invigorate a Tax Name Taxonomy so that tax names are as codified as possible. This allow for a better display on Invoice Description, and allowed us to implement a smart name_search.

In the community PR, we have changed the Taxes names so that it's more clear for users, but we need to changed in the test too.

closes odoo/enterprise#37914

Task-id: 3052677
Related: odoo/odoo#114653
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>