[FIX] website_knowledge: corrects WebsiteKnowledge patch
This commit fixes a bug introduced by a change in the standardFieldProps
changed in 688986f888f2fe2371d58b74ded81315ba6bb353.

We need to remove the key *value* from the props of the
CopyClipboardCharField component, the Component in the template was forgotten
when transitioning to the new standardFieldProps.

task-3223259

closes odoo/enterprise#37916

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>