[FIX] appointment: fixed taking appointment at not available hours
Steps to reproduce:

- Install the Appointment and Website.
- Create a new Online Appointment with certain hours available.
- Now when we go to the frontend to book an appointment.
- Now, taking into account that our default timezone is
'Europe/Brussels', we can see that the available hours are from 10:00
to 12:00.
- Now we change the timezone any timezone we want to use, for example
'America/New_York', now we can see that the available hours are now
from 5:00 am to 7:00am.
- Then we select 5:00am, and we go to the next step, here we can see
at the right side the hour that we have selected and the timezone.
- Now we change our mind and we want to change the hour that we have
selected, so we go back to the previous step but using the browser
button to go back (or the usual mouse button).
- In the first step again, we can see that we have the timezone reseted
to 'Europe/Brussels' again, with it rights hours.
- Let's say we forgot to change the timezone again, and we just select
one of the hours available for the current displayed timezone.
- We go to the next step and we look carfully at the right side, we can
see what happened with the hour that we have selected, and the timezone.

Issue:

With the flow of going back using the browser, we don't keep the proper
timezone in the session, so when we go back to the first step, the
frontend is updated for the default timezone, but we never really change
the previous timezone, which make us in this case for example, to be
able to book an appointment for 10:00am for 'America/New_York' timezone,
which is not an hour that we have available in our appointment.

Solution:

The issue is caused by the browser "storing" the page and not calling
the controller again when we go back which leads to not updating the
`request.session.timezone` key. So what we do here is to prevent the
browser from storing the page, so when we go back, the controller is
called again and the `request.session.timezone` is updated.

opw-3116509

closes odoo/enterprise#39279

X-original-commit: 037bb3ab54f24fb66203048be352c1ebdc72e941
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>