[IMP] l10n_de_pos_res_cert: adapt for cross-order feature
Related to odoo/odoo#109216

This commit aims at adding the feature "cross-orders" of the PoS.
This commit is related to a change made in community
where the name of a function (_updateTableOrder) change into _updateOrder.
This commit also fix a bug in the test tour of the
l10n_mx_edi_pos module.

closes odoo/enterprise#37955

Related: odoo/upgrade#4393
Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>