[FIX] l10n_be_reports: remove lines with foreign taxes
Steps to reproduce:
- Create a partner with belgian country and VAT.
- Create a 1.000 EUR invoice with this partner using the "0% Non EU" tax.
- Fully reverse the invoice.
- Duplicate the original invoice and apply the domestic "21%" tax.
- Open the partner VAT listing.

Observed behavior:
The turnover column of the partner is 0.0 EUR and the VAT amount is 210 EUR.

Desired behavior:
The turnover column should be equal to 1.000 EUR.

The problem was coming from 'tax_report_line_49' tags that are used for both
domestic and foreign operations.
There was no way to distinguish them using tags alone.
Which is why we need to filter out the move lines having this "0% Non EU" tax.

opw-3188086

closes odoo/enterprise#38274

X-original-commit: 4b88f65f0824bc2c808b0f14ddc596a20f73a2ef
Related: odoo/odoo#115493
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>