[FIX] l10n_be_reports: remove lines with foreign taxes
Steps to reproduce:
- Create a partner with belgian country and VAT.
- Create a 1.000 EUR invoice with this partner using the "0% Non EU" tax.
- Fully reverse the invoice.
- Duplicate the original invoice and apply the domestic "21%" tax.
- Open the partner VAT listing.

Observed behavior:
The turnover column of the partner is 0.0 EUR and the VAT amount is 210 EUR.

Desired behavior:
The turnover column should be equal to 1.000 EUR.

The problem was coming from 'tax_report_line_49' tags that are used for both
domestic and foreign operations.
There was no way to distinguish them using tags alone.
Which is why we need to filter out the move lines having this "0% Non EU" tax.

opw-3188086

closes odoo/enterprise#38472

X-original-commit: d3f13f043d5d999c017c0adbbda47ce7145e8a46
Related: odoo/odoo#115973
Signed-off-by: John Laterre (jol) <jol@odoo.com>