[IMP] l10n_nl_report: taxes
We want to invigorate a Tax Name Taxonomy so that tax names are as codified as possible. This allow for a better display on Invoice Description, and allowed us to implement a smart name_search.

In the community PR, we change the taxes name so that it's more clear for the users. Doing so the tests need to be modified

closes odoo/enterprise#37953

Task-id: 3052677
Related: odoo/odoo#114705
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>