[FIX] pos_preparation_display: wrong stage mapping and no category issue
*:pos_preparation_display

Problem with categories:
Before, when a product without a category was sent to the display
preparation it caused a crash.

Now it is no longer possible to send products without a category to
the preparation display. Namely, products without a category are mainly
those linked to a discount or loyalty.

Problem with stages:
Previously, order_stages not belonging to a display preparation were
sent to it because of a bad filter in process_order.

Now the filter has been corrected and this problem is
no longer a problem.

closes odoo/enterprise#37957

Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>