[FIX] knowledge: underline all property field inputs
For the property fields of Knowledge, we would like to always underline
the property field inputs to indicate that they are editable. Currently,
some inputs get an underline only when the user focuses or hovers the
element. This commit will ensure that the underline will always be
visible. It will also add some css tweaks to ensure that all inputs take
the full width of the panel.

Related: https://github.com/odoo/enterprise/pull/37299
task-3221040

closes odoo/enterprise#37958

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>