[IMP] sale_subscription: split log into expansion and contraction
Before this comit, all MRR change were saved into a signed float field.
This commit add two fields to save the expansion and contraction of MRR and
split the event_type depending on the sign of the change of MRR.
The advantage is to allow group by event_type and differentiate the net changes
to the delta changes.

Taskid: 3151840
Part-of: odoo/enterprise#36390