[FIX] l10n_be_reports: EC Sales List: proper date scope and report audit
The report was being computed with the 'normal' date scope. Because of that, if an account configured to include its initial balance was used on some move line eligible by this report (because of a +44 on this line in Belgium, for example), it was always included, even if it was out of the selected date range. This was wrong; this report needs to apply its date bounds in a strict manner.

Also, the audit caret option was not working well: it called the Partner VAT Listing without any tag domain, and basically always showed all the move lines of the selected partner within the period. This commit decouples both reports and adds a specific audit function to the Belgian EC Sales List, applying the right domain depending on the tax tags related to the report line.

closes odoo/enterprise#37975

Signed-off-by: John Laterre (jol) <jol@odoo.com>