[FIX] account_followup: fix styling in boxed, bold layout
Steps to reproduce:

- Install account_followup
- Settings > Configure Document Layout > Boxed/Bold
- Accounting > Customer > Follow-up reports
- Pick one > Follow up > check Print checkbox > Print

Issue:

Boxed layout: empty cells will have overlapping borders which do not
match the outlines of the table
Bold layout: empty cells will be represented as a thin line instead
of a full cell

This happens because the template `cell_template_followup_report`
is applied on the lines instead of the cells.

Solution:

Apply the template on the cells instead.

opw-3110496

closes odoo/enterprise#37041

Signed-off-by: Quentin De Paoli <qdp@odoo.com>