[IMP] pos_preparation_display: swap payment and order button
This PR makes the enterprise modules compatible with the modification
in Odoo community.

Previously in PoS, the payment button was larger than the others and next to
the numpad. The other buttons were managed by a mixin and arranged
in a predefined order.

Now, the payment button only stays in place when the "order" button is
not present, currently it can only be present if the restaurant module
is installed and a printer configured.

When it is present, it takes the place of the payment button and the
latter is added to the mixins with the other button

closes odoo/enterprise#38476

X-original-commit: 6d53fcaa85367dcbdc3fd43e565b802afc14568b
Related: odoo/odoo#115988
Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>