[FIX] mrp_workorder: use the operation pdf for register lots
Quality checks of type "register consumed material" are
created automatically to register lot of track raw materials.

If a pdf has been set on the operation it won't be displayed on them and
thus may be lost if this is the only quality check compare to a bom
without tracked product.

This commit adds the operation worksheet document in the tablet for those
quality checks.

closes odoo/enterprise#38034

Opw: 3179472
X-original-commit: 709d246bec46fd17ed0c136226c232eda94cc0cd
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: William Henrotin (whe) <whe@odoo.com>