[REF] account_accountant: Bank reco form as owl component
The form side of the bank recon used the generic form view, which lead to performance issues, from the inability to control the rendering (especially deep renders that were triggered by the model)

With this PR, the form view is replaced with custom owl components.
The service is removed and communication between components is enabled by a shared state on the environment.
The components used have also been simplified and renamed.

Task-3193121

closes odoo/enterprise#40068

Related: odoo/upgrade#4776
Related: odoo/odoo#119185
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>