[FIX] hr_payroll_attendance: do not use the language
The consequence of the following two commits:
4ab33d300006215146ddbf7a19e754f9470f0966
9e677ea115c1494d029513428e3919f54d24b27f
Make the solution use the local language 'en_US'.
Unfortunately, some databases do not have this language installed.

The goal is to find the date of the check in.

For example, for an attendance issued in the timezone 'Brussel/Europe' on February 1st,
if the only language installed in the database is French
the function `_read_group` with `groupby=[..., 'check_in:day']` will return:
```py
check_in:day': '01 févr. 2023'
'__range': { 'check_in:day': { 'from': '2023-01-31 23:00:00', 'to': '2023-02-01 23:00:00'}}
```

As we cannot parse the date, we need to find '2023-02-01' with '2023-01-31 23:00:00'.
It is therefore necessary to use the timezone of the employee who encoded the attendance.

Important remark:
The strong assumption is that the employee himself encodes his attendances.
If this is not the case, it is a direct manager who is in the same timezone.

opw-3182713

closes odoo/enterprise#38434

X-original-commit: ccfa0a60f1702f464f4b4bdd7caf616d7ceff349
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>