[FIX] account_online_sync*: fix default date + issue in cron
default=fields.Datetime.now() is taking the date when the server started
we should use default=fields.Datetime.now instead.

When the cron is running and is having a validationError due to an error
when creating transactions for example (same sequence number already exists),
that error is catched with the except UserError and we do nothing with it.
Which means that on the next iteration, when the cron will try to commit,
the error will still be present and that will raise again. To prevent that,
we rollback when we encounter an error.

closes odoo/enterprise#38098

X-original-commit: 3a3f0153bd7826000422468ebe4a54026375c40f
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>
Signed-off-by: Cedric Snauwaert <csn@odoo.com>