[FIX] mrp: timer wrong values on list view
The value of the timer was wrong when timesheeting with multiple employees
at the same time. The problem was that the compute duration already
takes the current timesheets into account.
So the problem was that we added some time that was already in the duration.

task id : 3216277

Part-of: odoo/enterprise#38007