[FIX] l10n_cl_edi_stock: tax group template could be missing
As we add taxes on all demo products, we now see that we may
have an error if we have the module `l10n_cl_edi_stock`
installed but the localization not used, as we may not have
the tax group for a company.
We prevent it by not raising an error if it is missing.

task-3045026

closes odoo/enterprise#38163

Related: odoo/odoo#115222
Signed-off-by: William André (wan) <wan@odoo.com>