[FIX] sale_planning: see hours to plan
Steps:
Create a plannable service.
Create a SO with this product, and confirm it.
Click on "Hours to Plan" stat button.

Issue:
Traceback.

Cause:
If there is no planning slot linked to the SO,
SaleOrder.planning_initial_date is set to False.
Then, it is put in the context of the action we call.
Then JS tries to evaluate false as a date, and so it crashes.

Fix:
If there is no planning slot linked to the SO,
set SaleOrder.planning_initial_date to today.

task-3226804

closes odoo/enterprise#38031

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>