[IMP] hr_appraisal_survey: simplify survey form
In order to ease the creation of survey, we continue the simplification of the
option form in survey started in odoo/odoo#115297. With the introduction of the
field survey_type in survey_survey, the field is_appraisal is now redundant.
The use of is_appraisal field has been replaced by the use of survey_type
(== 'appraisal').

Although the survey_type appraisal exists, it isn't visible to the user because
that survey type can only be created and viewed in hr_appraisal module. Indeed,
we want to avoid that the user creates/modifies/deletes appraisal survey in the
survey module. So we don't display the survey type selection in hr_appraisal
module and in the survey module, the selection is displayed without the
appraisal option.

Technical notes:
- The appraisal survey type is added in the survey type selection field in
hr_appraisal_survey. To avoid displaying it in the survey module, we hide it in
css (because the radio widget doesn't support hiding an option).
- Although the replacement of the 'is_appraisal' field by the appraisal survey
type could greatly simplify the uninstall_hook, we have decided to keep all the
cleaning. Indeed, now that we don't add a new field to distinguish the
appraisal survey the domain remains correct after uninstalling
hr_appraisal_survey even if then the survey_type option 'appraisal' doesn't
exist anymore. Note that as one action domain is dependant of the added field
survey_id (survey_question_answer), even if we wanted, we couldn't get rid of
it completly.

Task-3061869

closes odoo/enterprise#38268

Related: odoo/upgrade#4464
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>