[FIX] planning: set alloc.h.=0 if alloc.%=0
Current behaviour:
When we set the allocation percentage on a planning shift, the
allocation hours is the reset to the full-time quantity for the time
period. (aka an implicit 100% allocation percentage)

Expected behaviour:
If we set the allocation percentage to 0, the
allocation hours should be 0 too.

Steps to reproduce:
- Install Planning
- Take a shift, set the allocation percentage to 0
- Observe that the allocation hours are the same as with 100%

Reason for the problem:
In the compute for the allocation hours, there was a fallback to 1
when calculating the ratio from the percentage.

Fix:
Remove the fallback, it should be safe, because the ratio is never
used in a division.

Affected versions:
- 14.0
- saas-15.2
- 16.0
- saas-16.1
- master

opw-3188631

closes odoo/enterprise#38085

X-original-commit: a8f2f6b8bba8c13cfa6b92bce60513d3ffec6aca
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>