[FIX] timesheet_grid: display sample data
Before this commit, when no data is found for the grid view in `My Timesheets` menu
then the sample data should be displayed but a traceback is occured instead of
displaying the grid view with the sample data. The reason is `daily` attribute
is not defined in `workingHours` data.

This commit fixes the issue by defining `daily` attribute as an empty object in the
object `workingHours` when we initialize the data because loading all the data for
the grid view.

Bug found during the development of task-2829389
task-3245969

X-original-commit: 09abd10679035ccf70667f05f9428302e866db7a
Part-of: odoo/enterprise#38910