[FIX] web_grid: avoid sticky the rows and sections when sample data
Before this commit, when no data is found and the sample data is used
with the content helper, the rows and sections title is in the same level
than the no content helper because the z-index has been increased to fix
the title when the user scrolls horizontally.

This commit avoids to fix the sections and rows title when the sample data
is used to avoid having a z-index on title cells.

part of task-3222154
task-3245969

X-original-commit: 4d404846379ba640d6dbecc900b9f105f57750a0
Part-of: odoo/enterprise#38910