[FIX] account_reports: add a visibile condition for $matchingChilds
There are currently a performance issue using the search bar of the
partner ledger report.  For a report of 1200 partners, with the setting
'Add totals below section' activated, the page will take
several seconds (depending on the number of filtered lines)
when using the search bar, before displaying the correct
lines.
The customer who reported the issue has 5000 partners, it's almost
unusable in his case.

The issue is that there's a function that is quite slow, which looks at
every children node that matches the search and make sure
their parent node is also matched (so you don't see children nodes
without their parent).
If all the lines are collapsed, normally there's no children and this
function shouldn't consume any time.  But, in v16, for the partner
ledger at least, there are one children node per line, which is not
displayed, which corresponds to the total of the line (if you activated
the setting 'Add totals below section').
-> There's an unnecessary time spent looking for parent nodes of
   children nodes that are not even displayed.

My fix is to select only visible children nodes.

Furthermore, there are some weird behaviors due to this function that
will be fixed.
For example, if you type 'total' in the search bar, you will see all the
parent nodes.  Because they all have a children node named 'Total%',
that you see only if you unfold the node.
Second example, if you unfold a node then collapse it, the html code for
the children is still present, so they will be considered for the search
but not visible (you will see their parent if one of the children
match).

opw-3134212

closes odoo/enterprise#38078

Signed-off-by: Quentin De Paoli <qdp@odoo.com>