[FIX] sale_subscription: fix renewal close reason condition
Currently, in sale_subscription when closing a SO automatically, a renewal
close reason can be added if a certain condition is met. However this
condition doesn't make a lot of sense. It will check that the parent SO has
a confirmed renewal SO with date_order bigger than the parent SO's end date.

In the case that the date_order (date that the SO was confirmed) is the same
as the end_date, it works well. However, for the case where the date_order is
after the SO's end_date, it doesn't make a lot of sense. Assuming that the
invoicing cron is run daily, the parent SO will be closed before the renewal
SO is confirmed. Therefore the renewal close reason won't be added. This commit
changes the condition for adding the renewal close reason to not take into
account dates. The only condition is that the parent SO has a confirmed renewal
SO.

task-3216015

closes odoo/enterprise#37769

Signed-off-by: Arnaud Joset <arj@odoo.com>