[FIX] appointment: not allow appointment outside of time slots
Current behaviour:
A portal user can change the date and time of the URL before posting
the form to take an appointment, forcing the site to book an
appointment outside of the specified time slots of the employee
who's booked.

Expected behaviour:
Even if a user act smarts and changes the URLs parameters,
the backend shouldn't validate such invalid appointment.

Steps to reproduce:
- Install Website and Appointments
- Go to the front-end of an appointment of your choice
- Click on a slot, edit the hour in the URL for example to 23h00
- Press Enter to load the page with the new URL
- Post the form
- Observe that we've booked an appointment for 23h00, shouldn't be possible

Reason for the problem:
There is not backend validation in the controller that the start and end date
of an appointment are actual valid slots. We were supposing they were correct
because the date intervals comes from the previous page whose dates were generated
with the available slots in mind.

Fix:
Add a validation in the controller to check if the dates are indeed viable slots.
Cons is that the slots need to be regenerated once more, which may have a small
impact on performance. An eventual future improvement maybe to cache the slots on
the session to keep them between page loads if the performance becomes a problem.

Affected versions:
- 14.0
- 15.0
- saas-15.2
- saas-15.3
- 16.0
- master

opw-3080297

closes odoo/enterprise#38962

X-original-commit: 7d199987c73d26272fe98df2e666452f9a733e86
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>