[FIX] account_reports: support translated fields in Profit and Loss
Current behavior: if new translated field is added via studio on account.move.line, traceback appears in profit and loss report when analytic accounting is activated. This is due to the fact that jsonb field should not be cast as other type.

This commit allows to cast translated fields as 'jsonb'

opw-3208813

closes odoo/enterprise#38674

X-original-commit: 4cf8213f88f5eee45382386a3389e51cb2b75f4b
Signed-off-by: William André (wan) <wan@odoo.com>