[FIX] account_predictive_bills: add opt-out predictions
In some cases it is beneficial to not have predictions done
on the different fields of a bill (e.g. for Italian EDIs the product and
taxes shouldn't be predicted).
`account_predictive_bills_disable_prediction` already allows to
disable such predictions. But this adds a more granular way to do it.
With control over the product, account and taxes predictions in an
opt-out manner (i.e. by default the predictions are done).

closes odoo/enterprise#38079

Related: odoo/odoo#115072
Signed-off-by: William André (wan) <wan@odoo.com>