[FIX] l10n_mn_reports: Correction BS
When checking if the BS is balanced with a script,it seems that some accounts were missing in the BS and some sign were reverse, this pr corrects that.

In this commit, I've removed an account_id.account_type part of the domain of the "report_line_balanceca4" because the account 1202 and 1203 are not account receivable, so they were not added to the BS which was wrong. Doing so I've not remove other account_id.account_type part since they were good.

closes odoo/enterprise#38673

Task-id: 3041738
X-original-commit: 28e3112b86e0ccb5680645f37d1dcfaaf8f39f67
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>