[FIX] account_asset: deferred revenue sign fix
The deferred revenues have an issues with the signs of the
amounts.
Created a deferred revenue from an invoice should result in
a positive Original Value, but it is now negative.
Changing this also means changing the way the depreciation move lines
are computed to inverse it again and keep the correct amounts (as it is
now if the def rev is properly configured)

Also fix a more general issue when making a deferred revenue/expense or
asset model from the account.account list. The pre-populated account on
these models were wrong (it was populating the account_asset_id and not
account_depreciation_id)

Also revert https://github.com/odoo/enterprise/commit/eb7edf11fc0b26ea76b90f1408f31f60d484ca9f

Task id # 3229301

closes odoo/enterprise#38158

Signed-off-by: William André (wan) <wan@odoo.com>