[IMP] knowledge: buffer external steps during Component rendering
Use the mechanism introduced in `OdooEditor` to buffer external steps when some
asynchronous rendering needs to be done (to prevent external steps occuring
inside the rendered result from being applied when the rendering is currently
ongoing).

Provide a callback for the editor to scan for Behaviors and return the promise
resolved when they are rendered. If the editor gets a promise, external steps
will be buffered until the resolution.

Task-2821374

Part-of: odoo/enterprise#33483