[FIX] industry_fsm: fix sign button visible issue
Description of the issue/feature this PR addresses:
In the website when any task is open, sign button is visible all the the time.
Sign button should only visible when field service option is enable from the
particular project form view, worksheet template is not empty, task is in sign
state and worksheet field in project setting is enabled.

Current behavior before PR:
Sign button is visible in all task in website.

Desired behavior after PR is merged:
Sign button is only visible when field service option is enable from the
project form view, worksheet template is not empty, task is in sign
state and worksheet field in project setting is enabled.

Fix:
Add t-if condition in sign button template to check whether field service is
selected and sign report button is visible or not in task form view. Sign button
is only visible when any worksheet template is selected and worksheet field in
project setting is enabled.

task-3217427

closes odoo/enterprise#38875

Related: odoo/odoo#116805
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>