[FIX] website_sale_renting: uses a chosen date
When the test was running near 22:50, almost 2 weeks before the time change day, the test env was
losing one hour exactly at the time of the return, thus making the return for the next day instead,
which wasn't unavailable as it should have been.
With this commit, we avoid using that specific hour.

closes odoo/enterprise#38254

X-original-commit: a49596624349f4af5ff106b4e70c1a50bd88a6ed
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>