[FIX] website_sale_renting: uses a chosen date
When the test was running near 22:50, almost 2 weeks before the time change day, the test env was
losing one hour exactly at the time of the return, thus making the return for the next day instead,
which wasn't unavailable as it should have been.
With this commit, we avoid using that specific hour.

closes odoo/enterprise#38188

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>