[FIX] timesheet_grid: no edit timesheet of proj. w/ timesheet disabled
Current behaviour:
You are able to edit in the grid view the timesheet on a project
that *had* the option "Timesheets" enabled, but doesn't anymore.

Expected behaviour:
You should be able to edit a timesheet time in the grid view when
the associated project has the setting disabled. Else it will create
a new entry for the task in the project, even tough the option has
been disabled.

Steps to reproduce:
- Install Project, Timesheets
- Create a project with Timesheets enabled, on a task log some hours
- Validate the timesheet in Timesheets
- Disable the option on the project for timesheets
- In Timesheets > Grid View > try to modify the time of the
  timesheet we just validated
- The changes goes through and you can see in the list view that a
  new timesheet has been created with the difference.

Reason for the problem:
The cell in the grid view is not read only in that case.

Fix:
We cannot put the cell read only, but in `adjust_grid` we raise an
UserError if `allow_timesheets` on the project of the line is disabled.

Affected versions:
- 16.0
- saas-16.1
- master

opw-3203241

closes odoo/enterprise#38160

X-original-commit: 3cbbf13
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>