[FIX] sign: ensure the creation of an attachment
Steps to reproduce:
- install 'hr_contract_salary'
(to show template contract configuration);
- go to 'Employees/Configuration/Contracts/Templates';
- create a new template;
- create a new "Contract Document Template" via the form view;
- upload a file;
- click on 'SAVE & CLOSE' button;

Issue:
A User Error is triggered.

Cause:
In the code:
`attachments = self.env['ir.attachment'].browse([vals.get('attachment_id') for vals in vals_list])`
returns `attachments = [False]`.
Therefore,
`self._check_pdf_data_validity(attachment.datas)` to check if the data `b''` is valid.
As this is not the case, a user error is triggered.

Solution:
You have to make sure that an attachment is created
when we do not pass in the `create_with_attachment_data` function before the `create` function.

closes odoo/enterprise#39294

X-original-commit: e56a22a0f70fe9f4888739a51b2e495f2f7c23ff
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>