[FIX] spreadsheet: give model to action_download_spreadsheet
Previously, we gave the exported data to the action_download_spreadsheet
action, and we created a new model based on the data. This was a problem
for data that was only exported for the xlsx in UI plugins, because this
wasn't in the exported data.

Fixed by giving the model to the action_download_spreadsheet action instead
of the data.

Odoo task 3231170

closes odoo/enterprise#42355

X-original-commit: 1982074cef25a74c7e3072226e332d7d135f8421
Related: odoo/odoo#124617
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Minne Adrien (adrm) <adrm@odoo.com>