[FIX] *_extract: fix upload cron crash on concurrent update error
The uploading cron could sometimes crash due to a concurrent update
error.
In that case, the other documents awaiting upload wouldn't be sent to
the extraction server before the next cron execution, leading to a long
delay for the user.

The previous code wasn't handling the error properly as the savepoint
wasn't rollbacked in that case.

Ticket #3207721

closes odoo/enterprise#38445

X-original-commit: 1f2ea40600d688b26c54773f72ac71af14965f14
Signed-off-by: Arthur Gossuin (goa) <goa@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>