[IMP] l10n_no_{reports, saft}: taxes
We want to invigorate a Tax Name Taxonomy so that tax names are as codified as possible. This allow for a better display on Invoice Description, and allowed us to implement a smart name_search.

In the related community PR we have changed tax names so those test need to be modified

closes odoo/enterprise#38271

Task-id: 3052677
Related: odoo/odoo#115480
Signed-off-by: John Laterre (jol) <jol@odoo.com>