[FIX] helpdesk_stock: be able to return deliveries with open backorders
Current behaviour:
When you have a SO with an undelivered backorder, you are unable to
return the already delivered products from the helpdesk's ticket
"return" button, you are met with an UserError that the deliveries
needs to be "Done" so they can be returned.

Expected behaviour:
You shouldn't be met with an UserError, and the user should be able
to return the the deliveries that are already made, even if they
have an undelivered backorder.

Steps to reproduce:
- Install Helpdesks, Sales, Inventory, helpdesk_stock
- Create a SO, sell 5x storable product, confirm
- Deliver only some of the quantity (ex: 3)
- Create a backorder
- Create a helpdesk ticket for that costumer we just made the so
- Click on the Return button
- We are met with an UserError

Reason for the problem:
When we open the wizard for the return, an onchange is triggered
with the picking_id chosen for the wizard, which happened to be the
"last" picking that is linked to the so, which happens to be the
backorder that is not done. The onchange has an assertion that
picking in done should be returnable.

Fix:
Filter the chosen picking for the wizard to select only picking in
the SO that are actually in the `suitable_picking_ids`.

Affected versions:
- saas-15.2
- 16.0
- saas-16.1
- saas-16.2
- master

opw-3192714

closes odoo/enterprise#39197

X-original-commit: eb7aa8cc9ede0b2183abe59ce506e26d98b6171c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>