[MOV] timesheet_grid: move timesheet grid setup in helper file
Before this commit, the timesheet grid setup functions cannot really
do to be override and patched.

This commit moves the functions to setup the timesheet grid in helper
file and create a class TimesheetGridSetupHelper to add those functions
as methods for this class to be able to override/patch the class.

task-2829389

Part-of: odoo/enterprise#39706