[FIX] appointment: fix onboarding copy to clipboard
Purpose
=======
Fix the traceback appearing when trying to
copy the link of the appointment onboarding on
Chrome version inferior to 110.0.5481.100.

Specification
=============
Fixed the issue by setting in the setTimeout async method
the modal closing and the page reload to avoid executing
them before the copy to clipboard. Reloading the page was
messing with the browser copy action.

Task-3223000

closes odoo/enterprise#38288

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>