[IMP] l10n_cl_{reports,edi,boletas}: taxes
We want to invigorate a Tax Name Taxonomy so that tax names are as codified as possible. This allow for a better display on Invoice Description, and allowed us to implement a smart name_search.

In the community PR, we changed the name of the taxes so we need to correct thoses tests

closes odoo/enterprise#38286

Task-id: 3052677
Related: odoo/odoo#115483
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>