[IMP] l10n_lt_saft: taxes
We want to invigorate a Tax Name Taxonomy so that tax names are as codified as possible. This allow for a better display on Invoice Description, and allowed us to implement a smart name_search.

In the community PR, we changed the name of the taxes so we need to correct thoses tests

closes odoo/enterprise#38285

Task-id: 3052677
Related: odoo/odoo#115487
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>