[IMP] sale_renting_crm: propagate lead user to rental quotation
In order to match the logic of sale_crm, we want the salesperson
set on the lead to be propagated on the quotation, so that the
reporting correctly accounts for them. The user set on the lead
is added to context for both actions leading to the rental
quotations. It will be prioritized as the salesperson on the
newly created quotation, if any. Note: we want to add it as a
default user_id only if set, otherwise user_id will be False
by default in the newly created quotation.

Before, the next salesperson in line inherited the quotation,
i.e. the one set on the partner_id. (For more details on order
of user_id based on the partner_id, check the _compute_user_id
method in the sale.order model)

TEST

A test is added in TestLead to ensure the order of priority
for salesperson on the rental order from the lead.

Task-2612207

closes odoo/enterprise#38302

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>