[FIX] documents_spreadsheet_crm: load correct measures
In the "MRR/NRR Pipeline Revenue Report (Monthly)" spreadsheet template,
Some pivot measures are not loaded.

The measures in the pivot definition are "recurring_revenue_monthly",
"recurring_revenue_monthly_prorated", "expected_revenue" and
"prorated_revenue".

But there's only two of them in the pivot context ("expected_revenue"
and "prorated_revenue")

It turns out the context has priority over search params[1]

This redundant measures should probably be cleaned. I'm not sure if
we really need the ones in the context (maybe for the "See records"
feature?). As a bug fix, for minimum risk, this commit simply adds the
missing measures in the context.

Note: this issue was reported by XPL while writing the user documentation.
It shows this template was never used by anyone. That's why I'm only
fixing it in 16.0 and not before. I don't want a huge diffs everywhere.
(in master, those json files should be prettified)

[1] https://github.com/odoo/odoo/blob/273b99adc34d478d62d107040eae1db3ca883f6e/addons/web/static/src/views/pivot/pivot_model.js#L686

closes odoo/enterprise#38305

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>