[FIX] web_gantt: Disable drag for Gantt group pills
Before this commit, the "fake" draggable feature was applied to gantt
group pills. This meant that dragging the very thin group pills (2 pixel
high), you had the same visual feedback as when dragging a pill locked
because of a start or end date overflowing the current view.

Now only non-consolidated pills (i.e. the normal pills mentioned above)
are assigned that feature.

Task 3058801

closes odoo/enterprise#38366

X-original-commit: 12abbd6f08a926bfd1aaab0a36428bbf8ed7763f
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>