[FIX] stock_barcode: Allow to create a line without serial
In backend it's allowed to create a new inventory line with a product
tracked by serial but without specify it (to only track at delivery
level).

However in barcode it's not possible via the button to create a new
line because the form view has a required on the form if the product
is tracked.

It's not risky to remove it because there is a warning message at
validation to ensure it's not a mistake that the serial has not been
set

closes odoo/enterprise#37281

X-original-commit: 3a9dd11
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>