[FIX] knowledge: correctly destroy/unmount apps
Versions:
16.0+

How to reproduce:
- open an article in Knowledge
- create a / item list (list view)
- remove it (i.e. with `backspace` key)
- resize the browser window

Current Behavior:
- traceback

Expected Behavior:
- no traceback

Fix:
Removing a "mounted" component from the dom without destroying it implies that
all its listeners are still active in the document, and it can cause crashes
like the one with list redimensionning. As such, we have to destroy any
component when it leaves the DOM, and reconstruct it when it comes back (i.e.
remove the list with backspace, and re-insert it with CTRL+Z (undo)).

Instead of using Editor events, the proposed strategy in this commit is to use a
Mutation Observer on the editable of the html field, because this way we are
independent from editor implementation changes, and we don't have to maintain an
ever growing list of editor event listeners in the Knowledge patch of the html
field.

As the MutationObserver handle mutations after they already occured (i.e. a
removed element is already out of the DOM), it is sometimes necessary to put the
anchor of a Behavior back in the DOM to destroy the App and its Components. It
is done by introducing a `d-none` element next to the editable.

A mutex is also introduced for the method "updateBehaviors", in order not to
have it run multiple times at once (if the first one is still "awaiting", a
mutation could cause another call to it in the meantime).

Task-3236156

closes odoo/enterprise#38345

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>